chore : usePlacesHandlers -> usePlacesFieldArray 이름 변경
- INITIAL_PLACE 상수 추가
- useFieldArray가 제공하는 append, remove, swap 같은 로우레벨 API를 감싸서 "장소 배열"이라는 특정 도메인에 맞게 추상화한 커스텀 훅
