+++ src/constants/logConstants.ts
@@ -1,0 +1,9 @@
+import { LogFormValues } from '@/types/log';
+
+export const INITIAL_PLACE: LogFormValues['places'][number] = {
+  placeName: '',
+  category: '',
+  location: '',
+  description: '',
+  placeImages: [],
+};
